+++ torch_geometric/nn/aggr/adaptive.py
@@ -13,9 +13,9 @@ from torch_geometric.utils import to_dense_batch
 
 class MLPAggregation(Aggregation):
     r"""Performs MLP aggregation in which the graph nodes are flattened to a
-    tensor and are processed by a multi-layer perceptron, as described in the
-    `"Graph Neural Networks with Adaptive Readouts"
-    <https://arxiv.org/abs/2211.04952>`_ paper.
+    single tensor representation for the entire batch and are then processed by
+    a multi-layer perceptron, as described in the `"Graph Neural Networks with
+    Adaptive Readouts" <https://arxiv.org/abs/2211.04952>`_ paper.
 
     .. warning::
         :class:`MLPAggregation` is not a permutation-invariant operator.
@@ -51,8 +51,8 @@ class MLPAggregation(Aggregation):
             ), ('Cannot specify MLP arguments if a custom MLP is provided.')
 
         if num_hidden_layers is not None:
-            assert num_hidden_layers > 0, (
-                'Number of hidden layers must be non-zero.')
+            assert num_hidden_layers >= 0, (
+                'Number of hidden layers must be non-negative.')
 
         self.in_channels = in_channels
         self.out_channels = out_channels
@@ -74,7 +74,7 @@ class MLPAggregation(Aggregation):
             ]
 
             # Hidden layers
-            for _ in range(self.num_hidden_layers - 1):
+            for _ in range(self.num_hidden_layers):
                 modules.append(
                     nn.Linear(in_features=self.hidden_dim,
                               out_features=self.hidden_dim))
@@ -145,11 +145,11 @@ class MAB(nn.Module):
 
         A = torch.softmax(
             Q_.bmm(K_.transpose(1, 2)) / math.sqrt(self.dim_V), 2)
-        O = torch.cat((Q_ + A.bmm(V_)).split(Q.size(0), 0), 2)
-        O = O if getattr(self, 'ln0', None) is None else self.ln0(O)
-        O = O + F.relu(self.fc_o(O))
-        O = O if getattr(self, 'ln1', None) is None else self.ln1(O)
-        return O
+        Out = torch.cat((Q_ + A.bmm(V_)).split(Q.size(0), 0), 2)
+        Out = Out if getattr(self, 'ln0', None) is None else self.ln0(Out)
+        Out = Out + F.relu(self.fc_o(Out))
+        Out = Out if getattr(self, 'ln1', None) is None else self.ln1(Out)
+        return Out
 
 
 class SAB(nn.Module):
@@ -164,13 +164,13 @@ class SAB(nn.Module):
 class ISAB(nn.Module):
     def __init__(self, dim_in, dim_out, num_heads, num_inds, ln=False):
         super(ISAB, self).__init__()
-        self.I = nn.Parameter(torch.Tensor(1, num_inds, dim_out))
-        nn.init.xavier_uniform_(self.I)
+        self.Induced = nn.Parameter(torch.Tensor(1, num_inds, dim_out))
+        nn.init.xavier_uniform_(self.Induced)
         self.mab0 = MAB(dim_out, dim_in, dim_out, num_heads, ln=ln)
         self.mab1 = MAB(dim_in, dim_out, dim_out, num_heads, ln=ln)
 
     def forward(self, X):
-        H = self.mab0(self.I.repeat(X.size(0), 1, 1), X)
+        H = self.mab0(self.Induced.repeat(X.size(0), 1, 1), X)
         return self.mab1(X, H)
 
 
@@ -211,10 +211,8 @@ class SetTransformerAggregation(Aggregation):
     r"""Performs Set Transformer aggregation in which the graph nodes are
     considered to be set elements processed by multi-head attention blocks, as
     described in the `"Graph Neural Networks with Adaptive Readouts"
-    <https://arxiv.org/abs/2211.04952>`_ paper.
-
-    .. warning::
-        :class:`MLPAggregation` is a permutation-invariant operator.
+    <https://arxiv.org/abs/2211.04952>`_ paper. SetTransformerAggregation is a
+    permutation-invariant operator.
 
     Args:
         max_num_nodes_in_dataset (int): Maximum number of nodes in a graph in
@@ -311,3 +309,94 @@ class GRUAggregation(Aggregation):
     def __repr__(self) -> str:
         return (f'{self.__class__.__name__}({self.in_channels}, '
                 f'{self.out_channels}, {self.num_layers})')
+
+
+class DeepSetsAggregation(Aggregation):
+    r"""Performs Deep Sets aggregation in which the graph nodes are transformed
+    by a multi-layer perceptron (MLP) :math:`\phi`, summed, and then
+    transformed by another MLP :math:`\rho`, as suggested in the `"Graph Neural
+    Networks with Adaptive Readouts" <https://arxiv.org/abs/2211.04952>`_ paper
+    . DeepSetsAggregation is a permutation-invariant operator.
+
+    Args:
+        max_num_nodes_in_dataset (int): Maximum number of nodes in a graph in
+            the dataset.
+        in_channels (int): Size of each input sample.
+        out_channels (int): Size of each output sample.
+        num_hidden_layers_phi (int): Number of hidden layers for the
+            :math:`\phi` MLP.
+        num_hidden_layers_rho (int): Number of hidden layers for the
+            :math:`\rho` MLP.
+        hidden_dim (int): Dimensions used for the hidden layers of the MLPs.
+
+    """
+    def __init__(self, max_num_nodes_in_dataset: int, in_channels: int,
+                 out_channels: int, num_hidden_layers_phi: int,
+                 num_hidden_layers_rho: int, hidden_dim: int):
+        super().__init__()
+
+        assert num_hidden_layers_phi >= 0, (
+            'Number of phi hidden layers must be non-negative.')
+        assert num_hidden_layers_rho >= 0, (
+            'Number of rho hidden layers must be non-negative.')
+
+        self.max_num_nodes_in_dataset = max_num_nodes_in_dataset
+        self.in_channels = in_channels
+        self.out_channels = out_channels
+        self.hidden_dim = hidden_dim
+        self.num_hidden_layers_phi = num_hidden_layers_phi
+        self.num_hidden_layers_rho = num_hidden_layers_rho
+
+        # Input layers
+        self.phi = [
+            nn.Linear(in_features=self.in_channels,
+                      out_features=self.hidden_dim),
+            nn.ReLU(),
+        ]
+        self.rho = []
+
+        for _ in range(self.num_hidden_layers_phi):
+            self.phi.append(
+                nn.Linear(in_features=self.hidden_dim,
+                          out_features=self.hidden_dim))
+            self.phi.append(nn.ReLU())
+
+        for _ in range(self.num_hidden_layers_rho):
+            self.rho.append(
+                nn.Linear(in_features=self.hidden_dim,
+                          out_features=self.hidden_dim))
+            self.rho.append(nn.ReLU())
+
+        # Output layer
+        self.rho.append(
+            nn.Linear(in_features=self.hidden_dim,
+                      out_features=self.out_channels))
+
+        self.phi = nn.Sequential(*self.phi)
+        self.rho = nn.Sequential(*self.rho)
+
+        self.reset_parameters()
+
+    def reset_parameters(self):
+        def reset(network):
+            for module in network.children():
+                if hasattr(module, 'reset_parameters'):
+                    module.reset_parameters()
+
+        reset(self.phi)
+        reset(self.rho)
+
+    def forward(self, x: Tensor, index: Optional[Tensor] = None,
+                ptr: Optional[Tensor] = None, dim_size: Optional[int] = None,
+                dim: int = -2) -> Tensor:
+        x = self.phi(x)
+        x, _ = to_dense_batch(x, index, fill_value=0,
+                              max_num_nodes=self.max_num_nodes_in_dataset)
+        x = torch.sum(x, dim=1)
+        return self.rho(x)
+
+    def __repr__(self) -> str:
+        return (f'{self.__class__.__name__}({self.max_num_nodes_in_dataset}, '
+                f'{self.in_channels}, {self.out_channels}, '
+                f'{self.num_hidden_layers_phi}, {self.num_hidden_layers_rho}, '
+                f'{self.hidden_dim})')